Fjernet circular dependency
