+++ src/ducks/behandlingsgrunnlag/selectors.js
@@ -5,7 +5,6 @@ import * as Utils from '../../utils';
 import { OrganisasjonSelectors } from '../organisasjoner';
 
 import { PersonSelectors } from '../personer';
-import { behandlingerSelectors } from '../behandlinger';
 
 import MKV from '../../melosyskodeverk';
 
@@ -122,29 +121,6 @@ export const SelvstendigNaringsvirksomhetSelector = createSelector(
   SelvstendigArbeidForetakSelector,
   OrganisasjonSelectors.organisasjonerSelector,
   (selvstendigForetak = [], organisasjoner) => organisasjoner.filter(organisasjon => selvstendigForetak.find(foretak => foretak.orgnr === organisasjon.orgnr))
-);
-
-export const AlleVirksomheterSelector = createSelector(
-  state => behandlingerSelectors.ArbeidsforholdeneSelector(state),
-  state => SelvstendigNaringsvirksomhetSelector(state),
-  state => ForetakUtlandSelector(state),
-  state => EkstraArbeidsgivereVirksomhetSelector(state),
-  (arbeidsforholdene, selvstendigNaringsvirksomhet, foretakUtland, ekstraArbeidsgivere) => {
-    const virksomhetsListe = [];
-    arbeidsforholdene.forEach(arbeidsforhold => {
-      virksomhetsListe.push({ kode: arbeidsforhold.arbeidsgiver.orgnr, term: arbeidsforhold.arbeidsgiver.navn });
-    });
-    selvstendigNaringsvirksomhet.forEach(selvstendigForetak => {
-      virksomhetsListe.push({ kode: selvstendigForetak.orgnr, term: selvstendigForetak.navn });
-    });
-    foretakUtland.forEach(foretak => {
-      virksomhetsListe.push({ kode: foretak.uuid, term: foretak.navn });
-    });
-    ekstraArbeidsgivere.forEach(ekstraArbeidsgiver => {
-      virksomhetsListe.push({ kode: ekstraArbeidsgiver.orgnr, term: ekstraArbeidsgiver.navn });
-    });
-    return virksomhetsListe;
-  }
 );
 
 export const OppholdUtlandSelector = createSelector(
